@@ -34,28 +34,30 @@
     "build-storybook": "build-storybook"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.8.3",
-    "@babel/preset-env": "^7.8.3",
-    "@babel/preset-react": "^7.8.3",
-    "@material-ui/core": "^4.8.3",
-    "@storybook/addon-actions": "^5.3.3",
-    "@storybook/addon-links": "^5.3.3",
-    "@storybook/addons": "^5.3.3",
-    "@storybook/react": "^5.3.3",
-    "@testing-library/react": "^9.4.0",
-    "babel-loader": "^8.0.6",
-    "coveralls": "^3.0.9",
-    "jest": "^24.9.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "rollup": "^1.29.0",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-copy": "^3.3.0"
+    "@babel/core": "^7.15.5",
+    "@babel/plugin-transform-runtime": "^7.15.0",
+    "@babel/preset-env": "^7.15.6",
+    "@babel/preset-react": "^7.14.5",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.3.0",
+    "@mui/material": "^5.0.1",
+    "@storybook/addon-actions": "^6.3.8",
+    "@storybook/addon-links": "^6.3.8",
+    "@storybook/addons": "^6.3.8",
+    "@storybook/react": "^6.3.8",
+    "@testing-library/react": "^12.1.0",
+    "babel-loader": "^8.2.2",
+    "coveralls": "^3.1.1",
+    "jest": "^27.2.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "rollup": "^2.57.0",
+    "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-copy": "^3.4.0"
   },
   "peerDependencies": {
-    "react": "^16.8.0 || ^17.0.0",
-    "react-dom": "^16.8.0 || ^17.0.0",
-    "@material-ui/core": ">= 3.0.0"
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
+    "@mui/material": ">= 5.0.0"
   }
 }
